+++ src/core/lazy/CHANGELOG.md
@@ -9,6 +9,24 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.74.3 (2022-01-13)
+
+#### :house: Internal
+
+* Improved type inferring
+
+## v3.74.2 (2022-01-12)
+
+#### :bug: Bug Fix
+
+* Don't drop all registered actions after the first invoking
+
+## v3.74.1 (2022-01-12)
+
+#### :bug: Bug Fix
+
+* Now function hooks can be overridden
+
 ## v3.65.0 (2021-11-18)
 
 #### :rocket: New Feature