+++ src/core/object/watch/CHANGELOG.md
@@ -9,6 +9,18 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.73.3 (2022-01-10)
+
+#### :bug: Bug Fix
+
+* Fixed emitting of mutation events produced via `Object.defineProperty`
+
+## v3.73.0 (2021-12-14)
+
+#### :bug: Bug Fix
+
+* Now all properties that are added via `Object.defineProperty` can be handled via a proxy-based watcher
+
 ## v3.72.0 (2021-12-10)
 
 #### :bug: Bug Fix